requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.0